Rename action files to keep the project convention
